ks-112 improved readability /put RisValidationException in file with class/ fixed settings to give a correct message if slt is not correct or is missinh
