+++ tests/test_api_kount.py
@@ -9,7 +9,7 @@ import logging
 import unittest
 from json_test import example_data_products
 from kount.client import Client
-from kount.util.ris_validation_exception import RisValidationException
+from kount.ris_validator import RisValidationException
 from kount.settings import RAISE_ERRORS
 import inittest
